LayoutTests/inspector/editor/text-editor-token-at-position-expected.txt

Test editor tokenAtTextPosition method.


Running: testHighlightedText
Line: function foo(a, b) {
Column #0 (char 'f') - token: {"startColumn":0,"endColumn":7,"type":"javascript-keyword"}
Column #10 (char 'o') - token: {"startColumn":9,"endColumn":11,"type":"javascript-ident"}
Column #13 (char 'a') - token: {"startColumn":13,"endColumn":13,"type":"javascript-ident"}
Line:     var f = /.a/.test(a); /*
Column #8 (char 'f') - token: {"startColumn":8,"endColumn":8,"type":"javascript-ident"}
Column #14 (char 'a') - token: {"startColumn":12,"endColumn":15,"type":"javascript-regexp"}
Column #19 (char 's') - token: {"startColumn":17,"endColumn":20,"type":"javascript-ident"}
Line: this is a comment */
Column #0 (char 't') - token: {"startColumn":0,"endColumn":19,"type":"javascript-comment"}
Column #5 (char 'i') - token: {"startColumn":0,"endColumn":19,"type":"javascript-comment"}
Line:     return f + "looongword";
Column #5 (char 'e') - token: {"startColumn":4,"endColumn":9,"type":"javascript-keyword"}
Column #20 (char 'n') - token: {"startColumn":15,"endColumn":26,"type":"javascript-string"}

Running: testUnhighlightedText
Line: function foo(a, b) {
Column #0 (char 'f') - token: {"startColumn":0,"endColumn":7,"type":"javascript-keyword"}
Column #10 (char 'o') - token: {"startColumn":9,"endColumn":11,"type":"javascript-ident"}
Column #13 (char 'a') - token: {"startColumn":13,"endColumn":13,"type":"javascript-ident"}
Line:     var f = /.a/.test(a); /*
Column #8 (char 'f') - token: {"startColumn":8,"endColumn":8,"type":"javascript-ident"}
Column #14 (char 'a') - token: {"startColumn":12,"endColumn":15,"type":"javascript-regexp"}
Column #19 (char 's') - token: {"startColumn":17,"endColumn":20,"type":"javascript-ident"}
Line: this is a comment */
Column #0 (char 't') - token: {"startColumn":0,"endColumn":3,"type":"javascript-keyword"}
Column #5 (char 'i') - token: {"startColumn":5,"endColumn":6,"type":"javascript-ident"}
Line:     return f + "looongword";
Column #5 (char 'e') - token: {"startColumn":4,"endColumn":9,"type":"javascript-keyword"}
Column #20 (char 'n') - token: {"startColumn":15,"endColumn":26,"type":"javascript-string"}
